requirements.txt
Flask>=3.0.0
requests>=2.31.0
pillow